Read the CLI's wait window from one place
The authorization page tells the person how long the CLI keeps its listener
open, and the CLI enforces that window. Two copies of the number drift: the
page quoted a quarter of an hour while the listener waited half of one.

Both sides now read CLI_AUTH_TIMEOUT_MS from runtime-common, which both
packages already depend on, and the page renders it through the same
describeDuration the CLI prints — so the sentence a person reads and the
deadline they are racing cannot disagree.

Also states what the page does to the browser's own session accurately, now
that one of its three paths ends that session locally.
